Refactor _find_product_page into BaseSite
Add the SEARCH_REDIRECTED_TEXT class attribute to the BaseSite class.
This attribute can be set if a site redirects to the products page when
only one search result exists and there is unique text on the product
page to indicate this has occured.

Make the _find_product_page method in the BaseSite class a concrete
method instead of an abstract method, combining the FedcoSeeds and
BotanicalInterests methods.